修复输入不能用的bug

+++ yabause/src/port/qt/ui/UIControllerSetting.cpp
@@ -159,7 +159,7 @@ void UIControllerSetting::loadPadSettings()
 {
 	Settings* settings = QtYabause::settings();
 
-	foreach ( const u8& name, mNames.keys() )
+	foreach ( const uint& name, mNames.keys() )
 	{
 		mPadKey = name;
 		const QString settingsKey = QString( UIPortManager::mSettingsKey )
